Bug fix to 2 unit tests for COUNT() which were broken by the recent IV refactor.  The tests should have been expecting an xsd:long value, not an xsd:int value.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/TERMS_REFACTOR_BRANCH@5092 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/test/com/bigdata/bop/rdf/aggregate/TestCOUNT.java
@@ -108,7 +108,7 @@ public class TestCOUNT extends TestCase2 {
         for (IBindingSet bs : data) {
             op.get(bs);
         }
-        assertEquals(new XSDNumericIV(data.length), op.done());
+        assertEquals(new XSDNumericIV((long) data.length), op.done());
 
     }
 
@@ -164,7 +164,7 @@ public class TestCOUNT extends TestCase2 {
         for (IBindingSet bs : data) {
             op.get(bs);
         }
-        assertEquals(new XSDNumericIV(data.length), op.done());
+        assertEquals(new XSDNumericIV((long) data.length), op.done());
 
     }
